requirements.txt
Django==3.2.12
django_environ==0.9.0
djangorestframework==3.14.0
environ==1.0
PyMySQL==1.0.2
